requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
plotly>=5.17.0
psycopg2-binary>=2.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
